Update for Terraform 0.12 support
This changes the syntax of the entire project to be compatible with
Terraform 0.12.

There are some fixes that just clear out deprecation warnings, and 1 fix
that actually fixes an error that occurs because Terraform 0.12 kind of
changed how `compact` and interpolation works when doing

```
aliases = ["${compact(...)}"]
```

vs

```
aliases = compact(...)
```

This is the `aws_cloudfront_distribution.website`'s `aliases` variable
specifically.